+++ crates/gossamer-runtime/src/c_abi/symbol_table.rs
@@ -1815,6 +1815,14 @@ pub fn runtime_symbol_addrs() -> Vec<(&'static str, *const u8)> {
             crate::c_abi::gos_rt_iter_all_ptr as *const u8,
         ),
         (
+            "gos_rt_iter_any_ptr",
+            crate::c_abi::gos_rt_iter_any_ptr as *const u8,
+        ),
+        (
+            "gos_rt_iter_any_f64",
+            crate::c_abi::gos_rt_iter_any_f64 as *const u8,
+        ),
+        (
             "gos_rt_iter_any_i64",
             crate::c_abi::gos_rt_iter_any_i64 as *const u8,
         ),
@@ -3565,6 +3573,10 @@ pub fn runtime_symbol_addrs() -> Vec<(&'static str, *const u8)> {
         (
             "gos_rt_iter_position_i64",
             crate::c_abi::gos_rt_iter_position_i64 as *const u8,
+        ),
+        (
+            "gos_rt_iter_position_ptr",
+            crate::c_abi::gos_rt_iter_position_ptr as *const u8,
         ),
         (
             "gos_rt_iter_product_by_i64",